prototype change to use 32-bit value for the previous RPC format change.

+++ slashd/mdsio.h
@@ -90,7 +90,7 @@ struct mdsio_ops {
 	int	(*mio_fsync)(int, const struct slash_creds *, int, void *);
 	int	(*mio_getattr)(int, mdsio_fid_t, void *, const struct slash_creds *, struct srt_stat *);
 	int	(*mio_link)(int, mdsio_fid_t, mdsio_fid_t, const char *, const struct slash_creds *, sl_log_update_t);
-	int	(*mio_lookup)(int, mdsio_fid_t, const char *, mdsio_fid_t *, const struct slash_creds *, struct srt_stat *, int64_t *);
+	int	(*mio_lookup)(int, mdsio_fid_t, const char *, mdsio_fid_t *, const struct slash_creds *, struct srt_stat *, uint32_t *);
 	int	(*mio_lookup_slfid)(int, slfid_t, const struct slash_creds *, struct srt_stat *, mdsio_fid_t *);
 	int	(*mio_mkdir)(int, mdsio_fid_t, const char *, const struct srt_stat *, int, int, struct srt_stat *, mdsio_fid_t *,
 			sl_log_update_t, sl_getslfid_cb_t, slfid_t);
